feat(hooks): add useQueryExecution hook to execute queries
